Collisions Fixed

+++ Mamausa-Jump/Motor2D/j1Entity.h
@@ -40,7 +40,7 @@ public:
 	virtual bool Load(pugi::xml_node&) { return true; };
 	virtual bool Save(pugi::xml_node&) { return true; };
 
-	virtual void Draw(SDL_Rect r, bool flip = false, int x = 0, int y = 0);
+	virtual void Draw(SDL_Rect r, bool flip = false, float x = 0, float y = 0);
 	virtual void OnCollision(Collider* c1, Collider* c2) {};
 
 	ENTITY_TYPES type;
